requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
qrcode==7.4.2
reportlab==4.0.7
pandas==2.1.1
openpyxl==3.1.2